+++ src/config.py
@@ -23,7 +23,8 @@ DEFAULT_THEME = "arc"  # Standardtheme für die GUI
 # GUI-Einstellungen
 DEFAULT_WINDOW_SIZE = "800x600"  # Standardgröße des Anwendungsfensters
 HIGHLIGHT_DURATION = 2000  # Dauer der Texthighlights in Millisekunden
-DEFAULT_FONT_SIZE = 10  # Standard-Textgröße
+DEFAULT_FONT_SIZE = 12  # Standard-Textgröße
+DEFAULT_FONT_FAMILY = "Nimbus Mono L"  # Die korrekte Schreibweise ist "Nimbus Mono L"
 
 # Audio-Einstellungen
 AUDIO_FORMAT = pyaudio.paInt16  # 16-bit int Sampling